requirements.txt
altgraph==0.17.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==41.0.7
greenlet==3.2.3
idna==3.10
mysql-connector-python==9.3.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
PyMySQL==1.1.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.6
python-dotenv==1.0.0
pywin32==310
pywin32-ctypes==0.2.3
requests==2.31.0
setuptools==80.9.0
SQLAlchemy==1.4.50
typing_extensions==4.14.0
urllib3==2.4.0
qrcode[pil]==7.4.2